feat: consolidate duplicate secret creation
